+++ mycodo/scripts/upgrade_post.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 #
 #  upgrade_post.sh - Extra commands to execute for the upgrade process.
-#                   Used as a way to provide additional commands to
-#                   execute that wouldn't be possible from the running
-#                   upgrade script.
+#                    Used as a way to provide additional commands to
+#                    execute that wouldn't be possible from the running
+#                    upgrade script.
 #
 #  Copyright (C) 2015  Kyle T. Gabriel
 #
@@ -34,28 +34,21 @@ INSTALL_DIRECTORY=$( cd "$( dirname "${BASH_SOURCE[0]}" )/../../" && pwd -P )
 cd ${INSTALL_DIRECTORY}
 
 ln -sf ${INSTALL_DIRECTORY} /var/www/mycodo
-cp -f ${INSTALL_DIRECTORY}/mycodo_flask_apache.conf /etc/apache2/sites-available/
+cp -f ${INSTALL_DIRECTORY}/install/mycodo_flask_apache.conf /etc/apache2/sites-available/
 
 if [ -f "$INSTALL_DIRECTORY/mycodo_flask/ssl_certs/cert.pem" ] && [ ! -d "$INSTALL_DIRECTORY/mycodo/mycodo_flask/ssl_certs/" ]; then
     mkdir -p ${INSTALL_DIRECTORY}/mycodo/mycodo_flask/ssl_certs/
     cp ${INSTALL_DIRECTORY}/mycodo_flask/ssl_certs/* ${INSTALL_DIRECTORY}/mycodo/mycodo_flask/ssl_certs/
 fi
 
-${INSTALL_DIRECTORY}/mycodo/scripts/upgrade_mycodo.sh upgrade-packages
+/bin/bash ${INSTALL_DIRECTORY}/mycodo/scripts/upgrade_mycodo_release.sh upgrade-packages
 
 printf "#### Enable mycodo service ####\n"
 rm -rf /etc/systemd/system/mycodo.service
 rm -rf /etc/systemd/system/multi-user.target.wants/mycodo.service
-systemctl enable ${INSTALL_DIRECTORY}/mycodo/scripts/mycodo.service
+systemctl enable ${INSTALL_DIRECTORY}/install/mycodo.service
 
-printf "#### Upgrade influxdb if out-of-date ####\n"
-INFLUX_VERSION=$(apt-cache policy influxdb | grep 'Installed' | gawk '{print $2}')
-if [ "$INFLUX_VERSION" != "1.1.0-1" ]; then
-    echo "Incorrect version of InfluxDB installed ($INFLUX_VERSION). Downloading and installing version 1.1.0."
-    wget https://dl.influxdata.com/influxdb/releases/influxdb_1.1.0_armhf.deb
-    dpkg -i influxdb_1.1.0_armhf.deb
-    rm -rf influxdb_1.1.0_armhf.deb
-fi
+/bin/bash ${INSTALL_DIRECTORY}/mycodo/scripts/upgrade_mycodo_release.sh upgrade-influxdb
 
 printf "#### Move ssl certificates directory if not in correct directory\n"
 if [ -d "$INSTALL_DIRECTORY/mycodo/frontend/ssl_certs" ]; then
@@ -63,7 +56,7 @@ if [ -d "$INSTALL_DIRECTORY/mycodo/frontend/ssl_certs" ]; then
 fi
 
 printf "#### Checking if python modules are up-to-date ####\n"
-pip install --upgrade -r ${INSTALL_DIRECTORY}/requirements.txt
+pip install --upgrade -r ${INSTALL_DIRECTORY}/install/requirements.txt
 
 printf "#### Upgrading database ####\n"
 cd ${INSTALL_DIRECTORY}/databases
@@ -73,7 +66,7 @@ printf "#### Removing statistics file ####\n"
 rm ${INSTALL_DIRECTORY}/databases/statistics.csv
 
 printf "#### Setting permissions ####\n"
-${INSTALL_DIRECTORY}/mycodo/scripts/upgrade_mycodo_release.sh initialize
+/bin/bash ${INSTALL_DIRECTORY}/mycodo/scripts/upgrade_mycodo_release.sh initialize
 
 printf "#### Starting Mycodo daemon and reloading Apache ####\n"
 service mycodo start